+++ client/src/main.jsx
@@ -7,7 +7,7 @@ import App from "./App.jsx";
 import Home from "./pages/Home";
 import Error from "./pages/Error";
 import Search from "./pages/Search.jsx";
-import CreatePost from "./pages/CreatePost.jsx";
+import CreatePost from "./components/CreatePost.jsx";
 import ViewPost from "./pages/ViewPost.jsx";
 
 const router = createBrowserRouter([
